Combine Multiple type commands with Single type
There exists code duplication and similar functionality between
markm, unmarkm, and deletem commands with their single type
counterparts. They should be removed to increase code quality
and abide by DRY principle.

Changes has been made so that mark, unmark, and delete commands
can handle mass operations.
