+++ src/main/java/bob/enums/CommandType.java
@@ -4,5 +4,5 @@ package bob.enums;
  * Enumerations for types of commands available to be executed.
  */
 public enum CommandType {
-    BYE, LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, FIND, INVALID, MARKM, UNMARKM, DELETEM
+    BYE, LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, FIND, INVALID
 }
